project4-image-captioner/requirements.txt
torch>=2.0.0
transformers>=4.35.0
fastapi>=0.100.0
uvicorn>=0.20.0
pillow>=10.0.0
python-multipart>=0.0.6
httpx>=0.24.0
